requirements.txt
beautifulsoup4==4.6.0
boto==2.48.0
boto3==1.7.40
botocore==1.10.40
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.6
django-heroku==0.3.1
django-storages==1.6.6
djangorestframework==3.8.2
docutils==0.14
gunicorn==19.8.1
idna==2.7
jmespath==0.9.3
pipenv==2018.5.18
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.0
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
